Reverted to 32 bit ints, warning flag tweak.

+++ kernel.c
@@ -214,10 +214,10 @@ R lam1(Clos this, WORD _) {
 }
 
 R ping(Actor self, WORD q, Clos then) {
-    self->state[0] = (WORD)((int64_t)self->state[0] + 1);
-    int64_t j = (int64_t)self->state[0]*(int64_t)q;
+    self->state[0] = (WORD)((int)self->state[0] + 1);
+    int j = (int)self->state[0]*(int)q;
     if (j % 100000 == 0)
-        printf("Ping %ld\n", j);
+        printf("Ping %d\n", j);
     return (R){RCONT, CLOS3(lam1,self,q,then), None};
 }
 
@@ -226,9 +226,9 @@ R ping1(Clos this, WORD th) {
 }
 
 R pong(Actor self, WORD n, WORD q, Clos then) {
-    int64_t j = (int64_t)n*(int64_t)q;
+    int j = (int)n*(int)q;
     if (j % 100000 == 0)
-        printf("     %ld Pong\n", j);
+        printf("     %d Pong\n", j);
     ASYNC(self, CLOS2(ping1,self,q));
     return (R){RCONT, then, None};
 }
